zsh: Tweak prompt a bit more
With an FQDN that looks like host.xyz.example.com, this prompt will
show the host as "host.xyz", which is helpful if you have multiple
terminals open to various hosts with identical hostnames but differing
values of xyz. Effectively, just strip off the domain and tld from the
FQDN.

The above also applies for any subdomains beyond just 2 levels; those
are also included.

Also handles the case of host.local, stripping the .local.

The logic was already here, but force it to use whatever the FQDN was
when the shell was started. Before, it sorta depended on how the system
hostname was configured and subsequently stuffed into $HOST.

Finally, if a VPN connection or whatever changes the full domain/search
portion of the FQDN, now I can just run `prompt_read_host` to reread it.